Move permission filter to a more logical location
